+++ modules/plantType.js
@@ -1,13 +1,14 @@
 module.exports = {
-  carrot: { type: 'normal', seed: 'carrot', age: 7 },
-  potato: { type: 'normal', seed: 'potato', age: 7 },
-  beetroot: { type: 'normal', seed: 'beetroot_seeds', age: 3 },
-  wheat: { type: 'normal', seed: 'wheat_seeds', age: 7 },
-  melon: { type: 'melon', seed: 'melon_seeds' },
-  pumpkin: { type: 'melon', seed: 'pumpkin_seeds' },
-  oak_sapling: { type: 'tree', seed: 'oak_sapling' },
-  birch_sapling: { type: 'tree', seed: 'birch_sapling' },
-  jungle_sapling: { type: 'tree', seed: 'jungle_sapling' },
-  acacia_sapling: { type: 'tree', seed: 'acacia_sapling' },
-  dark_oak_sapling: { type: 'tree', seed: 'dark_oak_sapling' }
+  carrot: { plantName: 'carrots', type: 'normal', seed: 'carrot', age: 7 },
+  potato: { plantName: 'potatoes', type: 'normal', seed: 'potato', age: 7 },
+  beetroot: { plantName: 'beetroots', type: 'normal', seed: 'beetroot_seeds', age: 3 },
+  wheat: { plantName: 'wheat', type: 'normal', seed: 'wheat_seeds', age: 7 },
+  sweet_berries: { plantName: 'sweet_berry_bush', type: 'sweet_berries', seed: 'sweet_berries', age: 3 },
+  melon: { plantName: 'melon', type: 'melon', seed: 'melon_seeds', age: null },
+  pumpkin: { plantName: 'pumpkin', type: 'melon', seed: 'pumpkin_seeds', age: null },
+  oak_sapling: { plantName: '', type: 'tree', seed: 'oak_sapling', age: null },
+  birch_sapling: { plantName: '', type: 'tree', seed: 'birch_sapling', age: null },
+  jungle_sapling: { plantName: '', type: 'tree', seed: 'jungle_sapling', age: null },
+  acacia_sapling: { plantName: '', type: 'tree', seed: 'acacia_sapling', age: null },
+  dark_oak_sapling: { plantName: '', type: 'tree', seed: 'dark_oak_sapling', age: null }
 }