+++ docs/history.md
@@ -2,6 +2,9 @@
 * Realease a usable version
 * Fixed vec3 package
 * Some servers request a login or custom actions, added "custom start bot" do it your self
+* Fixed movements over "Wheat"
+* Fixed minor bugs
+* Adding sweet berries
 
 ## 0.1.9
 # Added in farmerjob new plants